+++ Changelog.md
@@ -7,9 +7,10 @@ Breaking Changes
 
 Features
 --------
+- Pass span initial references.
 - Pass span start time.
-- Span stores start time.
 - Span stores finish time.
+- Span stores start time.
 
 
 Version 0.1.1 (2018-01-16)